+++ tests/integration/long/test_ssl.py
@@ -34,6 +34,7 @@ SERVER_TRUSTSTORE_PATH = "tests/integration/long/ssl/.truststore"
 # Client specific keys/certs
 CLIENT_CA_CERTS = 'tests/integration/long/ssl/cassandra.pem'
 DRIVER_KEYFILE = "tests/integration/long/ssl/driver.key"
+DRIVER_KEYFILE_ENCRYPTED = "tests/integration/long/ssl/driver_encrypted.key"
 DRIVER_CERTFILE = "tests/integration/long/ssl/driver.pem"
 DRIVER_CERTFILE_BAD = "tests/integration/long/ssl/python_driver_bad.pem"
 
@@ -48,10 +49,11 @@ else:
     verify_certs = {'cert_reqs': ssl.CERT_REQUIRED,
                     'check_hostname': True}
 
+
 def setup_cluster_ssl(client_auth=False):
     """
     We need some custom setup for this module. This will start the ccm cluster with basic
-    ssl connectivity, and client authenticiation if needed.
+    ssl connectivity, and client authentication if needed.
     """
 
     use_single_node(start=False)
@@ -78,14 +80,18 @@ def setup_cluster_ssl(client_auth=False):
     ccm_cluster.start(wait_for_binary_proto=True, wait_other_notice=True)
 
 
-def validate_ssl_options(ssl_options):
+def validate_ssl_options(**kwargs):
+        ssl_options = kwargs.get('ssl_options', None)
+        ssl_context = kwargs.get('ssl_context', None)
+
         # find absolute path to client CA_CERTS
         tries = 0
         while True:
             if tries > 5:
                 raise RuntimeError("Failed to connect to SSL cluster after 5 attempts")
             try:
-                cluster = Cluster(protocol_version=PROTOCOL_VERSION, ssl_options=ssl_options)
+                cluster = Cluster(protocol_version=PROTOCOL_VERSION,
+                                  ssl_options=ssl_options, ssl_context=ssl_context)
                 session = cluster.connect(wait_for_all_pools=True)
                 break
             except Exception:
@@ -237,7 +243,7 @@ class SSLConnectionAuthTests(unittest.TestCase):
                        'ssl_version': ssl_version,
                        'keyfile': abs_driver_keyfile,
                        'certfile': abs_driver_certfile}
-        validate_ssl_options(ssl_options)
+        validate_ssl_options(ssl_options=ssl_options)
 
     def test_can_connect_with_ssl_client_auth_host_name(self):
         """
@@ -265,7 +271,7 @@ class SSLConnectionAuthTests(unittest.TestCase):
                        'certfile': abs_driver_certfile}
         ssl_options.update(verify_certs)
 
-        validate_ssl_options(ssl_options)
+        validate_ssl_options(ssl_options=ssl_options)
 
     def test_cannot_connect_without_client_auth(self):
         """
@@ -315,3 +321,50 @@ class SSLConnectionAuthTests(unittest.TestCase):
         with self.assertRaises(NoHostAvailable) as context:
             cluster.connect()
         cluster.shutdown()
+
+
+class SSLConnectionWithSSLContextTests(unittest.TestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        setup_cluster_ssl()
+
+    @classmethod
+    def tearDownClass(cls):
+        ccm_cluster = get_cluster()
+        ccm_cluster.stop()
+        remove_cluster()
+
+    def test_can_connect_with_sslcontext_certificate(self):
+        """
+        Test to validate that we are able to connect to a cluster using a SSLContext.
+
+        @since 3.17.0
+        @jira_ticket PYTHON-995
+        @expected_result The client can connect via SSL and preform some basic operations
+
+        @test_category connection:ssl
+        """
+        abs_path_ca_cert_path = os.path.abspath(CLIENT_CA_CERTS)
+        ssl_context = ssl.SSLContext(ssl_version)
+        ssl_context.load_verify_locations(abs_path_ca_cert_path)
+        validate_ssl_options(ssl_context=ssl_context)
+
+    def test_can_connect_with_ssl_client_auth_password_private_key(self):
+        """
+        Identical test to SSLConnectionAuthTests.test_can_connect_with_ssl_client_auth,
+        the only difference is that the DRIVER_KEYFILE is encrypted with a password.
+
+        @jira_ticket PYTHON-995.
+
+        @since 3.17.0
+        @expected_result The client can connect via SSL and preform some basic operations
+        @test_category connection:ssl
+        """
+        abs_driver_keyfile = os.path.abspath(DRIVER_KEYFILE_ENCRYPTED)
+        abs_driver_certfile = os.path.abspath(DRIVER_CERTFILE)
+        ssl_context = ssl.SSLContext(ssl_version)
+        ssl_context.load_cert_chain(certfile=abs_driver_certfile,
+                                    keyfile=abs_driver_keyfile,
+                                    password='cassandra')
+        validate_ssl_options(ssl_context=ssl_context)